Refactor: 예외처리 삭자
